feat(refactor): moved docs to client/src
