+++ anchor/client/Cargo.toml
@@ -24,6 +24,7 @@ http_api = { workspace = true }
 http_metrics = { workspace = true }
 hyper = { workspace = true }
 keygen = { workspace = true }
+keysplit = { workspace = true }
 logging = { workspace = true }
 message_receiver = { workspace = true }
 message_sender = { workspace = true }
